+++ src/pages/Home.tsx
@@ -6,6 +6,7 @@ import Button from "@mui/material/Button";
 import PlayArrowRoundedIcon from "@mui/icons-material/PlayArrowRounded";
 import ArrowBackRoundedIcon from "@mui/icons-material/ArrowBackRounded";
 import Layout from "../components/Layout";
+import AdsenseScript from "../components/AdsenseScript";
 import LanguageSelector from "../components/LanguageSelector";
 import SnakeDemo from "../components/SnakeDemo";
 import { useLanguage } from "../i18n/LanguageContext";
@@ -43,6 +44,7 @@ export default function Home() {
 
   return (
     <Layout showFooter>
+      <AdsenseScript />
       {fromHub && (
         <Box
           component="a"